Split Voltage into two forms.
I feel that the Rows is the less effective in this case. Both are such
that no read will ever be more than one average out at any point in
time. It seems that the rows gives odd results, in spite of only being
different in the manner that it reads. Sure. Alternating seems to get
fantastic results.